refactor(electron/renderer): moves Sidebar component directory
