+++ src/electron/renderer/features/sidebar/components/Sidebar/styles.ts
