+++ src/electron/renderer/features/sidebar/index.ts
@@ -1,0 +1,1 @@
+export * from './components/Sidebar';
